refactor: reorganise code in cli
