+++ packages/cli/bin/barnard59.js
@@ -9,7 +9,8 @@ import { NodeSDK } from '@opentelemetry/sdk-node'
 import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions'
 import { BatchSpanProcessor } from '@opentelemetry/tracing'
 
-import { Option, Command } from 'commander'
+import { Command } from 'commander'
+import * as otelOptions from '../lib/cli/otelOptions.js'
 
 const sdk = new NodeSDK({
   // Automatic detection is disabled, see comment below
@@ -43,22 +44,11 @@ const onError = async err => {
   // the same file, but we need to figure out what exporter is being used
   // before starting the SDK and loading any other code.
   const program = new Command()
-  const otelExporterOpt = new Option('--otel-traces-exporter <exporter>', 'OpenTelemetry Traces exporter to use')
-    .choices(['otlp', 'none'])
-    .default('none')
-  program.addOption(otelExporterOpt)
-  const otelMetricsOpt = new Option('--otel-metrics-exporter <exporter>', 'OpenTelemetry Metrics exporter to use')
-    .choices(['otlp', 'none'])
-    .default('none')
-  program.addOption(otelMetricsOpt)
-  const otelMetricsIntervalOpt = new Option('--otel-metrics-interval <milliseconds>', 'Export Metrics interval')
-    .argParser(value => Number.parseInt(value, 10))
-    .default(10000)
-  program.addOption(otelMetricsIntervalOpt)
-  const otelDebugOpt = new Option('--otel-debug <level>', 'Enable OpenTelemetry console diagnostic output')
-    .choices([...Object.keys(DiagLogLevel)].filter(opt => isNaN(Number.parseInt(opt, 10))))
-    .default('ERROR')
-  program.addOption(otelDebugOpt)
+
+  program.addOption(otelOptions.tracesExporter)
+  program.addOption(otelOptions.metricsExporter)
+  program.addOption(otelOptions.metricsInterval)
+  program.addOption(otelOptions.debug)
 
   // Command#parseOptions() does not handle --help or run anything, which fits
   // well for this use case. The options used here are then passed to the
@@ -98,8 +88,8 @@ const onError = async err => {
 
   // Dynamically import the rest once the SDK started to ensure
   // monkey-patching was done properly
-  const { run } = await import('../lib/cli.js')
-  await run([otelExporterOpt, otelMetricsOpt, otelDebugOpt, otelMetricsIntervalOpt])
+  const { default: run } = await import('../lib/cli.js')
+  await run()
   await sdk.shutdown()
 })().catch(onError)
