+++ packages/cli/lib/cli/commonOptions.js
@@ -1,0 +1,23 @@
+import { Option } from 'commander'
+
+function setVariable(str, all) {
+  let [key, value] = str.split('=', 2)
+
+  if (typeof value === 'undefined') {
+    value = process.env[key]
+  }
+
+  return all.set(key, value)
+}
+
+export const variable = new Option('--variable <name=value>', 'variable key value pairs')
+  .default(new Map())
+  .argParser(setVariable)
+
+export const variableAll = new Option('--variable-all', 'Import all environment variables')
+
+export const verbose = new Option('-v, --verbose', 'enable diagnostic console output')
+  .default(0)
+  .argParser((v, total) => ++total)
+
+export const enableBufferMonitor = new Option('--enable-buffer-monitor', 'enable histogram of buffer usage')
